changed to explicit map reduce function (#177)
* changed order of pages in side nav. changed the llm langchain map reduce to invoke map reduce explicitly for understanding the concept

* changed the example text

* changed to use pre-built image from public ECR since the image several GBs and not rworkshop setting is not reliable to build and upload since large image

* Update cdk/examples/generative_ai_service/web-app/pages/2_llm_langchain.py


---------
