+++ .gitignore
@@ -75,3 +75,5 @@ __pycache__/
 
 # Customized ENV files
 .env*
+.venv/
+output.json
