+++ cdk/examples/generative_ai_service/sample.env
@@ -19,6 +19,7 @@ az_count="3"
 
 # Gen AI Service Props
 container_name="web-container"
+container_image="public.ecr.aws/g4t0l7p7/sc-genai:latest"
 container_port="80"
 task_cpu="2048"
 task_memory="8192"
