+++ cdk/examples/generative_ai_service/web-app/pages/3_vector_search.py
