+++ cdk/examples/generative_ai_service/web-app/pages/4_image_generation.py
